eError now int not long to reduce overhead

+++ Source/Libraries/Util/include/eError.h
@@ -11,7 +11,7 @@
 //! http://forum.codecall.net/topic/56591-bit-fields-flags-tutorial-with-example/
 
 //! \brief generic eError enum for eError handling
-enum class eError : long
+enum class eError : int
 {
 // 	eError name		  eError bit value for flags
 	noErr			= 0x00000000,
