+++ Source/Libraries/Util/include/types.h
@@ -9,7 +9,7 @@
 #define _TYPES_H_
 
 //! \brief You can forward declare strongly typed enums!
-enum class eError : long;
+enum class eError : int;
 
 //! \brief all timer methods are in ms (milliseconds)
 typedef long ms;
